nhc_hobby_api/requirements.txt
dnspython==2.0
paho-mqtt==1.5